+++ src /types/User.ts
@@ -1,5 +1,10 @@
 export type User = {
     id: number;
     name: string;
-    email: string;
-};
+    email: string,
+    phone: string,
+    username: string,
+    address: {
+        city: string
+    };
+}